+++ packages/arb-bridge-eth/contracts/vm/IArbRollup.sol
@@ -22,7 +22,7 @@ interface IArbRollup {
         bytes32 _vmState,
         uint128 _gracePeriodTicks,
         uint128 _arbGasSpeedLimitPerTick,
-        uint32 _maxExecutionSteps,
+        uint64 _maxExecutionSteps,
         uint128 _stakeRequirement,
         address payable _owner,
         address _challengeFactoryAddress,
