+++ packages/arb-validator/ethbridge/executionchallenge/calls.go
@@ -19,7 +19,7 @@ func (_BisectionChallenge *BisectionChallengeTransactor) ChooseSegmentCall(ctx c
 	return callCheck(ctx, client, from, contractAddress, "chooseSegment", _segmentToChallenge, _proof, _bisectionRoot, _bisectionHash)
 }
 
-func (_ExecutionChallenge *ExecutionChallengeTransactor) BisectAssertionCall(ctx context.Context, client *ethclient.Client, from common.Address, contractAddress common.Address, _beforeInbox [32]byte, _timeBoundsBlocks [2]*big.Int, _machineHashes [][32]byte, _didInboxInsns []bool, _messageAccs [][32]byte, _logAccs [][32]byte, _gases []uint64, _totalSteps uint32) error {
+func (_ExecutionChallenge *ExecutionChallengeTransactor) BisectAssertionCall(ctx context.Context, client *ethclient.Client, from common.Address, contractAddress common.Address, _beforeInbox [32]byte, _timeBoundsBlocks [2]*big.Int, _machineHashes [][32]byte, _didInboxInsns []bool, _messageAccs [][32]byte, _logAccs [][32]byte, _gases []uint64, _totalSteps uint64) error {
 	return callCheck(ctx, client, from, contractAddress, "bisectAssertion", _beforeInbox, _timeBoundsBlocks, _machineHashes, _didInboxInsns, _messageAccs, _logAccs, _gases, _totalSteps)
 }
